+++ msdownloader.py
@@ -7,11 +7,13 @@ import datetime
 from dateutil.parser import parse
 import sys
 import argparse
+import pydash
+import collections
+
 
 # Function for retrieving the one-time bearer token from public website
-
-
 def get_ms_auth_token():
+
     url = "https://www.morningstar.co.uk/Common/funds/snapshot/SustenabilitySAL.aspx"
 
     querystring = {"Site": "uk", "FC": "F00000NOM7",
@@ -32,8 +34,6 @@ def get_ms_auth_token():
 
 # Function for downloading fund data based on the API URL scrapped from the site morningstar.co.uk
 # This function is actually not used.
-
-
 def get_ms_fund_data_failback(SAL, JWT, SEC):
     url = "https://www.us-api.morningstar.com/sal/sal-service/fund/esg/v1/" + SEC + "/data"
 
@@ -62,18 +62,39 @@ def get_ms_security_snapshot(SAL, JWT, ISIN):
     snapshot = json.loads(response.text)
     return snapshot
 
-# Function for downloading Fund/ETF closing prices from Screener API
-
-
-def get_ms_funds_prices(SAL, JWT, isin_list):
-    filters = ':'.join([str(isin) for isin in isin_list])
+def get_ms_security_data(SAL, JWT, term):
     url = "https://www.us-api.morningstar.com/ecint/v1/screener"
     querystring = {
         'outputType': 'json',
         'version': '1',
         'languageId': 'en-GB',
-        'securityDataPoints': 'ISIN,ClosePriceDate,ClosePrice,PriceCurrency',
-        'filters': 'ISIN:IN:' + filters
+        'securityDataPoints': 'SecId,Name,HoldingTypeId,Universe,PriceCurrency,ISIN',
+        'term': term
+    }
+    headers = {'x-sal-contenttype': SAL, 'Authorization': 'Bearer '+JWT, 'x-api-requestid': str(uuid.uuid4()),
+           'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.90 Safari/537.36'}
+    response = requests.request(
+        "GET", url, headers=headers, params=querystring)
+    data = json.loads(response.text)
+    return data
+
+# Function for downloading Fund/ETF closing prices from Screener API
+def get_ms_funds_prices(SAL, JWT, funds_list):
+    output_dict =[]
+    for d in funds_list:
+        d.setdefault('universe','FOEUR$$ALL')
+        output_dict.append(d['universe']) 
+    universe_list = '|'.join([str(universe) for universe in pydash.arrays.sorted_uniq(output_dict)])
+    filters_list = ':'.join([str(item['code']) for item in funds_list])
+
+    url = "https://www.us-api.morningstar.com/ecint/v1/screener"
+    querystring = {
+        'outputType': 'json',
+        'version': '1',
+        'languageId': 'en-GB',
+        'universeIds': universe_list,
+        'securityDataPoints': 'ISIN,ClosePriceDate,ClosePrice,PriceCurrency,Universe',
+        'filters': 'ISIN:IN:' + filters_list
     }
     headers = {'x-sal-contenttype': SAL, 'Authorization': 'Bearer '+JWT, 'x-api-requestid': str(uuid.uuid4()),
            'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.90 Safari/537.36'}
@@ -83,17 +104,21 @@ def get_ms_funds_prices(SAL, JWT, isin_list):
     return prices['rows']
 
 
-def get_ms_shares_prices(SAL, JWT, isin_list):
-    filters = ':'.join([str(isin) for isin in isin_list])
+def get_ms_shares_prices(SAL, JWT, shares_list):
+    output_dict =[]
+    for d in shares_list:
+        d.setdefault('universe','E0WWE$$ALL')
+        output_dict.append(d['universe']) 
+    universe_list = '|'.join([str(universe) for universe in pydash.arrays.sorted_uniq(output_dict)])
+    filters_list = ':'.join([str(item['code']) for item in shares_list])
     url = "https://www.us-api.morningstar.com/ecint/v1/screener"
     querystring = {
         'outputType': 'json',
         'version': '1',
         'languageId': 'en-GB',
-        'currencyId': 'BAS',
+        'universeIds': universe_list,
         'securityDataPoints': 'ISIN,ClosePriceDate,ClosePrice,PriceCurrency',
-        'universeIds': 'E0WWE$$ALL',
-        'filters': 'IsPrimary:EQ:True+ISIN:IN:' + filters
+        'filters': 'IsPrimary:EQ:True+ISIN:IN:' + filters_list
     }
     headers = {'x-sal-contenttype': SAL, 'Authorization': 'Bearer '+JWT, 'x-api-requestid': str(uuid.uuid4()),
            'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/77.0.3865.90 Safari/537.36'}
@@ -162,6 +187,7 @@ def read_args():
     group_input = arg_parser.add_mutually_exclusive_group(required=True)
     group_input.add_argument('-c', action='store', metavar='securities.yaml', default='securities.yaml', type=str, help='The YAML file with the list of securities')
     group_input.add_argument('-d', action='store', metavar='XY03ID03131ID', help='Dump info for a single security in JSON')
+    group_input.add_argument('-l', action='store', metavar='GB00B0XWNK36', help='Lookup for a security and print the simple data')
     arg_parser.add_argument('-x', action='store_true',  help='Force conversion from pence sterling GBX into pound sterling GBP')
     arg_parser.add_argument('-b', action='store_false',  help='Return beancount format instead of ledger/hledger')
     arg_parser.add_argument('-o', action='store', metavar='output.txt', type=str, help='The output file for the latest prices')
@@ -174,6 +200,8 @@ auth = get_ms_auth_token()
 
 if my_args.d is not None:
     print(json.dumps(get_ms_security_snapshot(auth[0],auth[1],my_args.d), indent=4, sort_keys=True))
+elif my_args.l is not None:
+    print(json.dumps(get_ms_security_data(auth[0],auth[1],my_args.l), indent=4, sort_keys=True))
 else :
     securities=read_securities(my_args.c)
     for group in securities.keys():
@@ -186,7 +214,7 @@ else :
         else:
             print('No match for ' + group)
 
-if my_args.o is not None:
-    save_prices(prices,my_args.o,my_args.w)
-else:
-    print_prices(prices)
+    if my_args.o is not None:
+        save_prices(prices,my_args.o,my_args.w)
+    else:
+        print_prices(prices)