+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
+- custom name for each security (not it's the ISIN)
+## [2.0.0] - 2021-01-18
+**BREAKING CHANGES**
+
+### Added
+- Support for custom universe
+- Lookup function for a security
 ## [1.3.0] - 2021-01-14
 ### Added
 - Support for shares